Standardize fact module error handling and align remaining dtypes
- Replace RuntimeError with SalesError in chunk_builder.py for
  pipeline-level failures: missing dimensions, capacity overflow,
  int32 overflow, FK mapping gaps, schema violations (7 sites)
- Replace RuntimeError with SalesError in returns_builder.py for
  config validation: lag days ordering, reason probs length (2 sites)
- Narrow returns reason_keys from int64 to int32 to match INT schema
  and eliminate redundant downstream casts
- Narrow inventory quantity_sold from int64 to int32 to match
  InventorySnapshot QuantitySold INT schema

+++ src/facts/inventory/micro_agg.py
@@ -88,5 +88,5 @@ def micro_aggregate_inventory(
         "store_key": g_store,
         "year": g_year.astype(np.int16),
         "month": g_month.astype(np.int8),
-        "quantity_sold": group_qty.astype(np.int64),
+        "quantity_sold": group_qty.astype(np.int32),
     }
